chore(NODE-3717): move tests p-z (#3098)
